refactor graph vs geo utils
