+++ osmnx/__init__.py
@@ -9,7 +9,6 @@ from .core import *
 from .downloader import *
 from .elevation import *
 from .footprints import *
-from .geo_utils import *
 from .plot import *
 from .pois import *
 from .projection import *
@@ -17,5 +16,7 @@ from .save_load import *
 from .simplify import *
 from .stats import *
 from .utils import *
+from .utils_geo import *
+from .utils_graph import *
 
 __version__ = '0.12.1'
