+++ osmnx/core.py
@@ -25,12 +25,12 @@ from .projection import project_geometry
 from .projection import project_gdf
 from .simplify import simplify_graph
 from .utils import log
-from .geo_utils import get_largest_component
-from .geo_utils import great_circle_vec
-from .geo_utils import get_nearest_node
-from .geo_utils import geocode
-from .geo_utils import count_streets_per_node
-from .geo_utils import overpass_json_from_file
+from .utils_graph import get_largest_component
+from .utils_geo import great_circle_vec
+from .utils_geo import get_nearest_node
+from .utils_geo import geocode
+from .utils_graph import count_streets_per_node
+from .save_load import overpass_json_from_file
 from .downloader import osm_polygon_download
 from .downloader import get_osm_filter
 from .downloader import overpass_request
