+++ osmnx/pois.py
@@ -14,7 +14,7 @@ from . import settings
 from .core import bbox_from_point
 from .core import gdf_from_place
 from .downloader import overpass_request
-from .geo_utils import geocode
+from .utils_geo import geocode
 from .utils import log
