+++ osmnx/stats.py
@@ -12,10 +12,10 @@ import numpy as np
 import pandas as pd
 
 from .simplify import clean_intersections
-from .geo_utils import get_largest_component
-from .geo_utils import great_circle_vec
-from .geo_utils import count_streets_per_node
-from .geo_utils import euclidean_dist_vec
+from .utils_graph import get_largest_component
+from .utils_geo import great_circle_vec
+from .utils_graph import count_streets_per_node
+from .utils_geo import euclidean_dist_vec
 from .utils import log
